dedicated class for RegEx processing
